Try to fix traces with lots of spans/events (#7886)
## Describe your changes
We aren't seeing all of the spans for these large vault proxy requests. I have a hypothesis that the otel collector might be running out of memory or something... So this PR removes a lot of the tracing events in these big events

Before, we would have a tracing::info for each fp_id that included each DI being decrypted.

Now, we only have that tracing::info when decrypting for a single fp_id (not in bulk decrypts). Then, I added some logs to the bulk decrypt utils that give us an idea of the range of fp_ids being handled in the bulk decrypt request and then a set of all of the DIs being decrypted across all fp_ids

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->