[FP-7135]: Flutter SDK: add proper access modifiers (#7916)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
## Why access modifiers are needed here:
In Flutter every file is considered a stand-alone library. That is
anything public on a file in the library can be accessed from the
integration code. The way to make separate files parts of the same
library is using `part` & `part of` (super weird structure, but read
more
[here](https://www.freecodecamp.org/news/how-to-use-part-in-dart/)). We
want tenants to have access to only:
- `footprint` object (not its class constructor)
- Configuration constructors
- `init` method on `footprint` object
Anything else should be internal to the library outside the scope of the
tenant. For instance, tenant doesn't and should know what the API base
URL is. This PR ensures that the tenant code can't access internal
methods, constructors, props, or variables.

### Feel free to add or remove any access modifiers as you deem
necessary

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->