Attempt to fix integration tests for ephemeral environment (#7889)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

See example failure:
https://github.com/onefootprint/monorepo/runs/20915820558

The problem is that the image is being built on the merge commit, but
the integration tests are running on the PR branch HEAD. I think this
might fix it -- I made the image build run on the PR commit rather than
a merge commit. I made a change to the cron job config to trigger the
ephemeral env build.

The integration tests now pass except for one that is marked as flaky
(it failed again after retrying).

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->