[FP-7144] Make identify flow stateful (#7910)
## Describe your changes
For the new "login with unverified email" flow, we're going to need to
start tracking some state even when the user is logged out. We're going
to require the user complete a kind of KBA, which we will store on their
backend session, before we allow them to log in with their unverified
email

To accomodate this, we'll issue an _identified_ (associated with a user)
but _unauthed_ (no scopes) token as soon as the user is identified in
`POST /hosted/identify`. This is probably the way I would have designed
this flow from the beginning, but we didn't have a mechanism for
"identified and unauthed" tokens then.

This identified token will be passed around to the other APIs in the
identify flow, and we'll store in this session the context on which KBAs
have been completed

## Caveats
We won't be able to deprecate the old version of the identify flow that
isn't stateful until we roll out these changes to all clients :(

## Test Plan
New integration test for the newer flow