Create Cloudfront 5xx alert instead of ALB 5xx alert (#7948)
## Describe your changes
Cloudfront is the outermost layer of our loadbalancing/firewalls. So,
all errors that occur as HTTP requests are routed through our stack will
be visible in Cloudfront.

We used to have alerts on the ALB target (ie the webserver) and ALB
itself throwing 5xxs. But those are only part of the picture since they
miss all errors that happen downstream of the ALB. Now, we'll also alert
on errors that originate at cloudfront to better catch cases like
https://onefootprint.slack.com/archives/C03DQ288QQM/p1706659855800229

## Caveats
This alert will fire when `HTTP 5xx Errors` fires as well. That other
alert is useful because you can immediately see the traces for the
requests that 5xxed. But maybe it's superfluous to have both of these
and we should just alert at cloudfront?

This also removes alerting on 5xxs from the nitro service ALB. But those
errors will bubble up into our application and should error loudly

## Test Plan
Ran this query in honeycomb:
https://ui.honeycomb.io/footprint-2e/environments/prod/datasets/cloudfront/result/aT21imNvnxD