Stop sending "Alert resolved" slack messages (#7949)
## Describe your changes
Small personal preference, but I am totally okay being told not to do
this -

Right now, most of our slack alerts give us a slack message when they
trigger AND when they are resolved. Just kind of makes things messy.
This changes our slack alerts to only fire when a trigger is true.

## Caveats
Honeycomb SUCKS - they make you choose the evaluation frequency of the
query AND the lookback period of the query. If the lookback period of
the query is shorter than the evaluation frequency, we'll dangerously
miss data. If the lookback period is longer, we can get double triggers.
And who knows what their SLA is on evaluating a trigger - what if they
miss the scheduled evaluation by more than (`lookback period` - `query
frequency`)? All of these things make me think our long term steady
state cannot really be on honeycomb.

For the purpose of this PR, this has kind of required me to change the
strategy such that the lookback period is fixed (right now, just 4m
since that's what it was for all queries before and changing it would
require changing thresholds). To prevent a lot of duplicate triggers
(which I am hypothesizing would happen), I decreased the slack alert
evaluation to every 3m instead of every 1m. We could also experiment
leaving the evaluation frequency to every 1m and see if duplicates fire,
I'm really not sure what honeycomb will do

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->