create helper struct to translate between what we've collected and rules (#7965)
## Describe your changes
For PoA step ups, we need to be able to step up after running rules on a collected document. For example, we could have a WF where:
* post KYC decisioning -> `address_does_not_match` -> StepUp(Identity)
* post Doc collection decisioning -> `document_ocr_does_not_match_address` -> StepUp(ProofOfAddress)
* post PoA collection -> `poa_collected` -> `ManualReview`

Currently we handle the post KYC/post doc collection rule dance via the `allow_stepup` param we pass to `evaluate_rules`, which is essentially "if we haven't collected an identity doc we can stepup". in the future, we need to have logic that expresses the available rule actions we can take, e.g. "which documents have we collected, and remove all rules that might step up to those documents from evaluation".

I think this is the most straightforward way to do this right now, maybe there's some better grouping of rules or stashing some sort of state on WF for the appropriate "rule group" to use in the next decisioning. BUt we're essentially doing this already so i think it makes sense

In order to orchestrate this we need to do 2 main things:
* Introduce an interface such that we can more granularly handle which specific RuleActions are applicable based on what docs have been collected
* Move WF into states based on the secondary step up rules (like move from decisioning -> doc collection if PoA is required)

This PR starts setting up the first

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->