[FP-7162] Handle non-HTTP-2xx responses in enclave client (#8002)
## Describe your changes
We've been seeing an HTTP error from what seems like the enclave proxy
client while trying to decrypt a handful of large VReses. It's hard to
find with our current visibility where the error is actually coming
from...

looks like an HTTP 413:
https://ui.honeycomb.io/footprint-2e/environments/prod/datasets/fpc-api/result/od7Es6JvpES/trace/Fu82ymUS5dY?fields[]=s_name&span=eefbe3210df698cb

The error we're actually seeing looks to be a decoding error from
reqwest, I think because we're trying to json decode the response as an
enclave response even if it's a non-2xx. So, this adds some special
handling of non-2xx responses to actually give us some insight into what
the error could be

I'm hoping this will tell me if the 413 is from the ALB or from the
enclave proxy itself...

## Caveats
A little spooky if there were real decryption responses that were_
non-2xx, but I don't think that's very likely

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->