Fix SQL random_string implementation (#8019)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
The generated string did not always have the expected length. Postgres `substring` is 1-indexed, so if `round(random() * 62) == 0` (0.8% of the time), then the substring step returns an empty string for that sampling round.

This probably doesn't matter that much and it's arguably a feature that ~40% of the time the string is shorter so we aren't implicitly making that an API contract. But in the interest of not stumbling on this in the future, I'm proposing the fix.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->