Simplify identify API when a user is found (#8022)
## Describe your changes
There are becoming a lot of fields that give context on the located user in `POST /hosted/identify`. while we're doing some surgery on the auth flow, this just nests all the user context under a `user` key in the response. The idea is that `user` will be populated when `user_found = true`, and `user` will be null when `user_found = false`.

## Caveats
This removes some fields from the original identify response - but i added them 3 days ago in #7909 and #7910, and nobody is using them yet

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->