[FP-7140] Track when a duplicate vault is created (#8040)
## Describe your changes
As part of the "login with unverified email" flowchart, if you cannot
log into your existing account AND if you don't already have a vault at
the provided tenant, we should allow you to make a new vault.

This starts to track when one of those new vaults is created by linking
the new vault to the old vault

## Caveats
We still don't yet block creating a duplicate vault yet - there are a
few cases that happen in prod today that i don't want to break until we
have better handling. For ex, it's possible you can be on the signup
challenge screen in the frontend and then make your vault in another
window, and our frontend won't yet properly direct you to the login
challenge. Once we build this, we can more thoroughly prevent creating
duplicate vaults and force you to log into your existing vault

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->