Add timeline event for stepups (#8072)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Adds new timeline event that is created when we have rules determine a
stepup action during a workflow. For now will just have the `kind` of
the requests but in followup PR's i add a link to rule_set_result_id and
it will have this as well so the FE can use that to lookup and display
the rules results that caused that stepup event.

## Caveats

<!--- Are there any caveats to this changeset? -->

Need to land a FE branch first to add support for this new timeline
event. (or confirm it won't barf display something funky) cc @zamandalee

## Test Plan

<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->