Fix instant app passkey (#8084)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
A tricky issue with android credentials with not a lot of documentation
out there.


https://github.com/onefootprint/monorepo/assets/54286440/c48c4099-3586-4166-8ec9-14f661eae350

### Before testing
You need a signed build to use the android credential manager. In order
to build and run a signed build in dev mode from android studio to test,
make sure to add the `mobile/apps/my/android/release.keystore`. And
update the debug `signingConfigs` in the instant app gradle as follows.
```
debug {
    storeFile file('../release.keystore')
    storePassword //use correct password
    keyAlias //use correct key
    keyPassword //use correct password
}
```
### To-do
- We change one prop `"requireResidentKey": true` on the requestJson
coming from BE. Maybe we want to ask BE to send the correct prop if it
work on iOS.
- We should use the android credential `version 1.2.0 (stable)`. The
`react-native-passkey` library uses an old version. We can commit the
changes in library gradle, but we should update the version in the
`react-native-passkey` gradle when we create the release `.aab` file.


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->