Add rule_set_result_id to DocReq (#8074)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Currently, when we execute rules and determine a StepUp action is to be
taken, we simply create DocReq's and move the user to the DocCollection
workflow state. There's not strong data model connection linking those
doc req(s) with the rule_set_result that caused it. Previously, this
wasn't a super duper big deal because a given workflow could only have 0
or 1 stepup rule_set_result and 1 final onboarding decision
rule_set_result. But now we (1) possible have 0, 1, or 2 stepup
rule_set_results (with the addition of proof of address stepup) and (2)
are going to be showing timeline events for stepups (PR's above this)
and those timeline events need a rule_set_result_id so the FE can
display the rule_set_result in the drawer that caused that stepup when
that event is clicked in the timeline.

I thiiiink also might be relatively easy to backfill this (modulo the
recent Alpaca stepup changes we just shipped which enormously complicate
things) so I'll take a stab at that too in a followup PR

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->