Add rule_set_result_id to obd timeline event (#8075)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

So FE can ping `GET
/entities/<entity_id>/rule_set_result/<rule_set_result_id>` (which
doesnt exist yet) to retrieve the correct rule_set_result for a given
OBD timeline event.

Currently, our timeline is a bit incorrect in that anytime you click on
an OBD event, it will show the rule_set_result from the _latest_
onboarding (but not necessarily the onboarding you clicked on if there
are multiple). So this fixes that and this same pattern will be used
with the StepUp timeline events

cc @zamandalee 

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->