[FP-7190] Migrate to more modern API for POST /hosted/identify (#8071)
## Describe your changes
The backend is migrating to a more modern API here that will make it easier to support more advanced features. namely, there's now a `user` envlope that contains all of the context on the located user, if any.

This PR updates the three hooks that use this API (and all their callers):
- One in the auth component
- One in the identify machine of the verify component
- One to support stepup in the confirm page of the collect KYC data machine in the verify component

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Unit tests, tested these flows locally