[FP-7200] Create challenge select screen in identify machine of auth component (#8090)
## Describe your changes
Okay, big novel:
- The identify flow is a set of APIs exposed by the backend that generally allows you to take an identifier (phone, email, or identified auth token) and turn it into an _authenticated_ auth token by challenging the user to either log in with an existing auth method OR create a new vault by verifying the phone/email. More context [here](https://www.notion.so/onefootprint/Identify-Flow-6a35783cc326488bb93bfd9af8ce4bbb)
- This logic has become fragmented around the frontend codebase. There's a version of the identify machine in the verify component, and then a nearly identicial identify machine in the auth component (called the auth machine). Then, on the "update login methods" feature, there's _another_, more modern version of the identify flow that only allows logging into existing vaults rather than creating new ones
- My dream is that there is only one set of code for the identify flow that is used in these 3 different places

The newest designs of the identify flow are in the "update login methods" machine, where we ask you to log in using any of the existing auth methods on your account. The old identify flow makes some pre-set decisions on which methods you can use to log in based on the kind of playbook you're logging into. This has caused problems especially with findigs' no-phone playbooks - there are users who _only_ have an email auth method, and they can't log in to their existing account and simply get stuck.

SO, this modernizes the normal identify flow (only in the auth component, will eventually port over to the verify component) to allow the user to log in using _any_ method on their profile. For the findigs no-phone case, we'll eventually allow a user to add a phone number to their vault after they log in with their email.

A few notes:
- This PR also starts storing pretty much the raw result from `POST /hosted/identify` in the machine context - previously, we would pick the response apart. But it's much more convenient to just automatically have all of the response fields available
- The biometric challenge doesn't require any ui elements - it happens in OS prompts. So, I put the logic to run the biometric challenge on the challenge select screen

## Caveats
I will soon update the "update login methods" component to use this single, more modern identify flow. Then, I will update the verify component to use this single modern identify flow. Especially as we start to make more changes to this identify flow, it will be really nice to have consistent behavior - for example, we're adding the ability to log in with your unverified email if you've lost access to your phone. It would be great to only have to implement this in one place

## Test Plan
It doesn't actually seem that this existing flow has unit tests... i'll add some before merging, but putting up for review with manual testing for now
'
Normal signup challenge:
https://github.com/onefootprint/monorepo/assets/21066341/2b1e68d9-e9ef-4763-9076-d1f2c618237b

User with a passkey and phone as verified login methods:
https://github.com/onefootprint/monorepo/assets/21066341/65a3f109-8abd-45c2-834e-8b47c01e3ca7

User with a passkey, phone, and email as verified login methods:
https://github.com/onefootprint/monorepo/assets/21066341/ecb53d90-4c7e-4b5d-9816-5439b9e682c0

User that had previously onboarded onto a no-phone playbook and only has an email:
https://github.com/onefootprint/monorepo/assets/21066341/e53f6547-f5e6-462b-8917-a499c4e96ff8

With bootstrapped phone + email:
https://github.com/onefootprint/monorepo/assets/21066341/9859d10a-70c4-442a-b320-cf598196369a