[FP-7174] Separate auth machine out from auth app as identify machine (#8109)
## Describe your changes
In the last few PRs, we've been modernizing the functionality of the
identify machine, as implemented inside the auth component. The goal is
to use this identify state machine in (1) the update login methods page
of the auth app and (2) the verify component.

This starts to separate out the implementation of the identify flow into
an "identify machine." Its interface is a single `<Identify />` react
component that takes in various properties (playbook key, auth token),
runs the user through the identify flow, and then calls an `onDone`
callback with the auth token returned by the identify machine.

This purely moves code around:
- `auth-container.tsx` -> split into `identify/identify.tsx` and
`identify-app.tsx`. The latter is all of the auth-app specific code
needed to invoke the identify machine, while the former contains the
nicely-packaged `<Identify />` component mentioned above
- `auth-router.tsx` -> `identify/components/router.tsx`
- all components that were used only in the identify machine have been
moved into `identify/components`

I'm still not quite sure where to put the general identify component
(https://onefootprint.slack.com/archives/C061Q3Y99UH/p1707439439319789),
but the goal is that this new `identify` folder can be moved pretty much
anywhere in our frontend monorepo

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally, unit tests