Add feature flag for identify machine rollout (#8179)
## Describe your changes
I'm using a LaunchDarkly flag from the backend, serialized on the
onboarding configuration, to control the rollout of a new feature on the
frontend. This is probably not the most canonical or recommended way to
do a feature rollout on the frontend, but it's the one I am most
familiar with 😛 i was getting some launch darkly errors and figured i'd
rather not learn how to do client-side launch darkly rollouts while
rolling this out. This also allows us to log on the backend whether the
user is going through the new/old identify flow, which may help us with
debugging

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->