Fix translation setup for apps that use idv (#8178)
## Describe your changes
I caught this because the builds on #8177 started failing due to this
change:
https://github.com/onefootprint/monorepo/pull/8177/files#diff-cdeb9808c64097cd361d407fac0cc921c930d2e1696a6b35ed99557a120375b1

It seems every app that uses `idv` needs to independently have these
definitions set up for all of the translation jsons in the `idv`
package. It's proving to be pretty laborious to add a new translation
json to a package - though admittedly this probably won't happen often
and `idv` is the most used package that will probably be the hardest to
add a translation json to

## Caveats
We should look into why `useTranslation('identify')` calls don't fail
typecheck while `<Trans ns="identify" ... />` invocations do fail to
build. It seems that when we add a new translation to a package used by
other apps that the `useTranslation` case could incorrectly pass
typecheck when the translations won't be available at runtime

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->