[FP-7201][idv] Start using the modern identify machine in a rollout (#8167)
## Describe your changes
Updates the API for the new `<Identify />` component to meet all cases
used by idv. I'm using a LaunchDarkly flag from the backend, serialized
on the onboarding configuration. This is probably not the most canonical
or recommended way to do a feature rollout on the frontend, but it's the
one I am most familiar with 😛 i was getting some launch darkly errors
and figured i'd rather not learn how to do client-side launch darkly
rollouts while rolling this out. This also allows us to log on the
backend whether the user is going through the new/old identify flow,
which may help us with debugging

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
UI tests for the old and new machines are practically identical, showing
feature parity.

And, here is some manual testing of 6 different cases of the identify
machine:

**Signing up for a new vault**


https://github.com/onefootprint/monorepo/assets/21066341/c303afd7-aeda-456f-be0a-d9890a1a532a


**Logging into an existing vault when only SMS challenge is available**


https://github.com/onefootprint/monorepo/assets/21066341/76aa6193-eaee-4bac-a09a-f9e4f070ae58

**Logging into an existing vault with SMS, biometric, and email
challenges available**
This is the first case that differs between the new and old machines -
the new machine has the ability to handle any kind of login method,
including email. We also have a more modernized challenge selector page


https://github.com/onefootprint/monorepo/assets/21066341/676623b3-c9cf-48f3-b233-ecd5ce90de36

**Logging into an existing vault that only has email as a login method,
on a playbook that requires phone**
The also differs in implementation - new machine handles this well and
allows the user to log in with the one login method they have. The old
machine shows an error.
NOTE: there's still more work to do to collect the phone when a playbook
requires it but the vault doesn't have a phone, but the first step was
allowing logging in at all in the first place


https://github.com/onefootprint/monorepo/assets/21066341/76f38683-7cfa-4e3b-a64b-ae330655f41b

**Signing up for a new vault on a no-phone playbook**


https://github.com/onefootprint/monorepo/assets/21066341/8f64218e-422a-4d4d-a71f-887e3c80cdb2

**Logging into an existing vault that only has an email on a no-phone
playbook**


https://github.com/onefootprint/monorepo/assets/21066341/31f159c4-2169-40fe-b78b-7a41190fd8ac