[FP-7188][identify] Register new phone as auth method if the playbook requires it (#8206)
## Describe your changes
After all of this work updating the identify machine, we're finally able
to solve one of the problems findigs has been reporting to us 🙂

Users who previously onboarded onto a no-phone playbook _only_ have an
email availalbe as a login method. If they ever try to one-click onto a
playbook that requires phone, they were previously bricked - they'd see
an ugly error message and be stuck.

With some recent improvements to the identify machine, these users are
now able to log in using their email if it's the only available login
method. And then, after this PR, we'll intelligently ask them to also
provide a phone number if the playbook requires one.

Also, it's very cool - we already had all the code to do this! I reused
the `UpdatePhone` component that we use in the "update login methods"
product

## Caveats
This could also be built as a machine outside of the identify machine,
or as a step of the onboarding requirements machine. I have no strong
opinion.

I just put it in the identify machine for now because it's also possible
in the future that auth playbooks will require a user to register
multiple auth methods.

## Test Plan
Added UI tests and state machine tests. Also tested locally:


https://github.com/onefootprint/monorepo/assets/21066341/cd1c575a-595e-4bec-81b9-d66110ec8a36