[identify][FP-7238] Skip past login challenge if token already has sufficient scopes (#8222)
## Describe your changes
This was a piece of functionality that existed in the old identify machine and i didn't think we'd need in the new one.

I forgot about a small use case where the backend will generate authenticated tokens (inherited auth). Nobody is using this yet, but this at least handles that case

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
UI tests, state machine tests, tested locally:

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/xncBfoFfDL8XmXxkoXEI/4bdde527-1789-4664-98d4-b9d5e5ca678e.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/xncBfoFfDL8XmXxkoXEI/4bdde527-1789-4664-98d4-b9d5e5ca678e.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/4bdde527-1789-4664-98d4-b9d5e5ca678e.mov">Untitled.mov</video>