[FP-7175] Soft error if a required auth method isn't provided (#8227)
## Describe your changes
In #8206, we started registering a phone number as a login cred if it is required by the playbook.

Eventually, we'll want to have a server-gated check that you're only onboarding to a playbook if you have the proper auth methods. For now, this just soft errors to make sure we won't be accidentally gating many onboardings from happening. In the future, we'll turn this into a hard error

## Caveats
This validate API is getting messy - it has mixed code to check for onboarding requirements (a very well-supported structure returned through `GET /hosted/onboarding/status`) and now to check "auth" requirements (a much less supported concept). I wouldn't be surprised if the identify machine ends up with its own mini `GET /hosted/auth/status` API or something similar to better codify requirements on auth/kyc playbooks that certain auth methods are provided

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->